+++ src/main/java/judamov/sipoh/repository/IStatusAvailabilityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IStatusAvailabilityRepository extends JpaRepository<StatusAvailability, Integer> {
+public interface IStatusAvailabilityRepository extends JpaRepository<StatusAvailability, Long> {
 }
