+++ src/main/java/judamov/sipoh/repository/IUserAreaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IUserAreaRepository extends JpaRepository<UserArea,Integer> {
+public interface IUserAreaRepository extends JpaRepository<UserArea,Long> {
 
 }
